env: add ReadV() API
Adds ReadV() methods to RWFile and RandomAccessFile that allows
reading data into multiple Slices in one call. The implementation
leverages the preadv system call when possible and simulates it
with pread calls when unavailable.

Additionally adds ReadV() methods to the block manager abstraction.
These methods will be used in KUDU-463 to support reading
checksums and block data in a single call.

Change-Id: Ib8f7a62c8363b40baa064d9e63be1ece506f1e48