KUDU-2412: Fix python client compilation in el6 environments
Identifies if the compiler supports __int128 and uses cython
preprocessors to ignore any code using __int128 if not. This
includes test code.

This is a bit of a hacky approach but I needed to
duplicate the entire cdef anytime I used a preprocessor
to remove methods inside of a `cdef extern`.
The preprocessor statement does not work inside of a
`cdef extern` statement and you can’t break out the cdef
staments to use them additively.

The cython preprocessors do not have a lot of
documentation. Perhaps there is a better approach I missed
that we could apply in a follow up.

See here for documentation:
http://cython.readthedocs.io/en/latest/src/userguide/language_basics.html#conditional-statements

Change-Id: Ibd93b57020b80597baae9c8d3e0434c46f7fc3d7