KUDU-2412: Fix python client compilation in el6 environments
This uses the C preprocessor to check the value of KUDU_INT128_SUPPORTED from the C++ headers
and defines a Cython macro PYKUDU_INT128_SUPPORTED.

PYKUDU_INT128_SUPPORTED is generated by
replacing the KUDU_INT128_SUPPORTED placeholder
in config.pxi.in to generate config.pxi which is then included
in client.pyx.

This is then used to conditionally generate the bits of Python code
that depend on int128-related methods.

See here for documentation:
http://cython.readthedocs.io/en/latest/src/userguide/language_basics.html#conditional-statements

Change-Id: Ibd93b57020b80597baae9c8d3e0434c46f7fc3d7