[java] Add equals and hashcode to Schema
Adds an equals and hashcode method to Schema
and PartitionSchema. Additionally fixes the equality
check on ColumnSchema to consider column attributes.

These methods are useul to quickly know that two
Schemas are exactly the same. We will use them in
the backup and restore testing, and they are likely
more useful than reference equality to users.

A few light tests were added to prove equality works.
Exhaustive testing was not done given that would be
a large amount of tests and the implementation is fairly
straightforward using Guava.

Change-Id: I0a6914b9b25df9e327ed6ae6bd61eac29c9c524c