[docker] Adjust Kudu image tagging
This patch defines a tagging scheme for the Kudu
docker images and adjusts the images to be a bit more
idiomatic for a docker user. A full description of the
tagging scheme can be found in docker-build.sh.

Example tags for the kudu image are:
   kudu:latest
   kudu:1.9.1
   kudu:1.9
   kudu:latest-centos7
   kudu:1.9.1-centos7
   kudu:1.9-centos7

I removed the kudu-master and kudu-tserver images in
favor of a single kudu image used for runtime applications.
I also adjusted the entrypoint script to be a bit more
flexible along with adding useful “help” output.

Change-Id: I4c96a3d5e1ca92022dc614c878f3853a4d99bcb8