[docs] Update the limitations page
This patch updates the scaling limitations to be a bit more
nuanced.  We have gotten feedback that users think the
limits listed are hard and fast rules and this change is
intented to be more clear about “just works” scale
vs some of the largest scales/configurations that have
been seen/reported.

A community survey will be conducted soon to adjust
these values further.

I also adjusted a few other out of date limitations.

Change-Id: Iafa4f7f3bd9e405f7ffc4e6cde48ec28e6e04081