[WIP] KUDU-2820: Generate JUnit XML reports for Java test run via dist-test
This patch adds a Junit `RunListener` that will generate a Jenkins
compliant Junit XML file in the `TEST_LOGDIR` directory. It also
adds the JunitXMLCore classwhich is a wrapper around the original
JunitCore, which was called by dist-test, to add the listener.

The dist-test scripts were updated to set `TEST_LOGDIR` to a
location similar to local java runs so that globing the test results
on Jenkins works in either local or dist-test. Additionally the
generated  dist-test isolate files now call JunitXMLCore instead
of JunitCore.

Change-Id: I7f65e0ef0a4c1ed2e79460f74bea6e725643c2ac