KUDU-1438: [java] Upgrade to Netty 4
This patch upgrades from Netty 3 to the latest Netty 4.
Below is a list of some of the changes required:
- Changed org.jboss.netty imports to io.netty
- Adjust shading to relocate io.netty
- Removed outdated reflection in Bytes.java
- Changed ChannelBuffer to ByteBuffer
- Changed OneToOneDecoder to BytesToMessageDecoder
- Changed OneToOneEncoder to MessageToByteEncoder
- Changed DecoderEmbedder to EmbeddedChannel
- Changed Negotiator and Connection to extend
  SimpleChannelInboundHandler instead of
  SimpleChannelUpstreamHandler
- Worked around non-sharable SslHandler
- Migrated from NioClientSocketChannelFactory to Netty Boostrap
- Removed boss threads and deprecated setter methods in the
  AsyncKuduClient
- Removed ShutdownThread in AsyncKuduClient
- Updated Negotiator tests to unwrap SSL messages
- Added build flag to run tests with paranoid level leak detection
- Suppressed errorprone FutureReturnValueIgnored warnings

I used YCSB to performance test this change and ensure there
isn’t a large performance regression. On an m4.2xlarge instance
I ran a test of 5 iterations of workload a, b, and c using 1M records
and 1M operations. The difference between runs was negligible
(< 1%) and within test variance.

Change-Id: Ic75bd15a982187039ff2e1510af9390d304f7626